+++ src/animation/utils/transitions.ts
@@ -3,18 +3,13 @@ import {
     PermissiveTransitionDefinition,
     ResolvedValueTarget,
 } from "../../types"
-import {
-    AnimationOptions,
-    KeyframeOptions,
-    Animatable,
-    animate,
-    inertia,
-} from "popmotion"
+import { AnimationOptions, animate, inertia } from "popmotion"
 import { secondsToMilliseconds } from "../../utils/time-conversion"
 import { isEasingArray, easingDefinitionToFunction } from "./easing"
 import { MotionValue } from "../../value"
 import { isAnimatable } from "./is-animatable"
 import { getDefaultTransition } from "./default-transitions"
+import { complex } from "style-value-types"
 import { warning } from "hey-listen"
 
 type StopAnimation = { stop: () => void }
@@ -42,7 +37,7 @@ export function isTransitionDefined({
 /**
  * Convert Framer Motion's Transition type into Popmotion-compatible options.
  */
-export function convertTransitionToAnimationOptions<T extends Animatable>({
+export function convertTransitionToAnimationOptions<T>({
     yoyo,
     loop,
     flip,
@@ -53,7 +48,7 @@ export function convertTransitionToAnimationOptions<T extends Animatable>({
     const options: AnimationOptions<T> = { ...transition }
 
     if (times) {
-        ;(options as KeyframeOptions<T>).offset = times
+        ;(options as any).offset = times
     }
 
     /**
@@ -150,11 +145,21 @@ function getAnimation(
     transition: PermissiveTransitionDefinition,
     onComplete: () => void
 ) {
-    const origin = value.get()
     const valueTransition =
         transition[key] || transition["default"] || transition
-    const isOriginAnimatable = isAnimatable(key, value.get())
+    let origin = value.get()
+
     const isTargetAnimatable = isAnimatable(key, target)
+
+    /**
+     * If we're trying to animate from "none", try and get an animatable version
+     * of the target. This could be improved to work both ways.
+     */
+    if (origin === "none" && isTargetAnimatable && typeof target === "string") {
+        origin = complex.getAnimatableNone(target as string)
+    }
+
+    const isOriginAnimatable = isAnimatable(key, origin)
 
     warning(
         isOriginAnimatable === isTargetAnimatable,