+++ src/render/dom/layout/__tests__/scale-correction.test.ts
@@ -14,7 +14,7 @@ describe("correctBoxShadow", () => {
     test("Correctly scales box shadow", () => {
         expect(
             correctBoxShadow(
-                "10px 10px 10px 10px #000",
+                "5px 10px 20px 40px #000",
                 {
                     x: { min: 0, max: 0 },
                     y: { min: 0, max: 0 },
@@ -28,7 +28,7 @@ describe("correctBoxShadow", () => {
                     y: 1,
                 }
             )
-        ).toBe("20px 20px 20px 20px rgba(0, 0, 0, 1)")
+        ).toBe("10px 20px 40px 80px rgba(0, 0, 0, 1)")
         expect(
             correctBoxShadow(
                 "10px 10px 10px 10px #000",
@@ -46,6 +46,47 @@ describe("correctBoxShadow", () => {
                 }
             )
         ).toBe("40px 40px 40px 40px rgba(0, 0, 0, 1)")
+
+        expect(
+            correctBoxShadow(
+                "10px 10px 10px 10px var(--token-c8953278-7b3-4177-a0fa-f3cda39afd50, rgba(0, 0, 0, 1))",
+                {
+                    x: { min: 0, max: 0 },
+                    y: { min: 0, max: 0 },
+                },
+                {
+                    x: { scale: 0.5, translate: 0, origin: 0, originPoint: 0 },
+                    y: { scale: 0.5, translate: 0, origin: 0, originPoint: 0 },
+                },
+                {
+                    x: 1,
+                    y: 1,
+                }
+            )
+        ).toBe(
+            "20px 20px 20px 20px var(--token-c8953278-7b3-4177-a0fa-f3cda39afd50, rgba(0, 0, 0, 1))"
+        )
+
+        // TODO: Only supports a single shadow
+        expect(
+            correctBoxShadow(
+                "10px 10px 10px 10px rgba(0, 0, 0, 1), 0px 0px 0px 4px var(--token-c8953278-7b3-4177-a0fa-f3cda39afd50, #333)",
+                {
+                    x: { min: 0, max: 0 },
+                    y: { min: 0, max: 0 },
+                },
+                {
+                    x: { scale: 0.5, translate: 0, origin: 0, originPoint: 0 },
+                    y: { scale: 0.5, translate: 0, origin: 0, originPoint: 0 },
+                },
+                {
+                    x: 1,
+                    y: 1,
+                }
+            )
+        ).toBe(
+            "10px 10px 10px 10px rgba(0, 0, 0, 1), 0px 0px 0px 4px var(--token-c8953278-7b3-4177-a0fa-f3cda39afd50, #333)"
+        )
     })
 })