+++ src/render/dom/layout/scale-correction.ts
@@ -1,5 +1,6 @@
 import { Point2D, Axis, AxisBox2D, BoxDelta } from "../../../types/geometry"
-import { complex } from "style-value-types"
+import { cssVariableRegex } from "../utils/css-variables-conversion"
+import { complex, px } from "style-value-types"
 import { mix } from "popmotion"
 
 type ScaleCorrection = (
@@ -34,10 +35,16 @@ export function correctBorderRadius(
     viewportBox: AxisBox2D
 ) {
     /**
-     * If latest is a string, we either presume it's already a percentage, in which case it'll
-     * already be stretched appropriately, or it's another value type which we don't support.
+     * If latest is a string, if it's a percentage we can return immediately as it's
+     * going to be stretched appropriately. Otherwise, if it's a pixel, convert it to a number.
      */
-    if (typeof latest !== "number") return latest
+    if (typeof latest === "string") {
+        if (px.test(latest)) {
+            latest = parseFloat(latest)
+        } else {
+            return latest
+        }
+    }
 
     /**
      * If latest is a number, it's a pixel value. We use the current viewportBox to calculate that
@@ -49,7 +56,7 @@ export function correctBorderRadius(
     return `${x}% ${y}%`
 }
 
-type BoxShadow = [string, number, number, number, number]
+const varToken = "_$css"
 
 export function correctBoxShadow(
     latest: string,
@@ -57,17 +64,35 @@ export function correctBoxShadow(
     delta: BoxDelta,
     treeScale: Point2D
 ) {
-    // GC Warning - this creates a function and object every frame
-    const shadow = complex.parse(latest) as BoxShadow
+    const original = latest
+
+    /**
+     * We need to first strip and store CSS variables from the string.
+     */
+    const containsCSSVariables = latest.includes("var(")
+    const cssVariables: string[] = []
+    if (containsCSSVariables) {
+        latest = latest.replace(cssVariableRegex, (match) => {
+            cssVariables.push(match)
+            return varToken
+        })
+    }
+
+    const shadow = complex.parse(latest)
+
+    // TODO: Doesn't support multiple shadows
+    if (shadow.length > 5) return original
+
     const template = complex.createTransformer(latest)
+    const offset = typeof shadow[0] !== "number" ? 1 : 0
 
     // Calculate the overall context scale
     const xScale = delta.x.scale * treeScale.x
     const yScale = delta.y.scale * treeScale.y
 
     // Scale x/y
-    shadow[1] /= xScale
-    shadow[2] /= yScale
+    ;(shadow[0 + offset] as number) /= xScale
+    ;(shadow[1 + offset] as number) /= yScale
 
     /**
      * Ideally we'd correct x and y scales individually, but because blur and
@@ -78,12 +103,25 @@ export function correctBoxShadow(
     const averageScale = mix(xScale, yScale, 0.5)
 
     // Blur
-    if (typeof shadow[3] === "number") shadow[3] /= averageScale
+    if (typeof shadow[2 + offset] === "number")
+        (shadow[2 + offset] as number) /= averageScale
 
     // Spread
-    if (typeof shadow[4] === "number") shadow[4] /= averageScale
+    if (typeof shadow[3 + offset] === "number")
+        (shadow[3 + offset] as number) /= averageScale
 
-    return template(shadow as any)
+    let output = template(shadow)
+
+    if (containsCSSVariables) {
+        let i = 0
+        output = output.replace(varToken, () => {
+            const cssVariable = cssVariables[i]
+            i++
+            return cssVariable
+        })
+    }
+
+    return output
 }
 
 const borderCorrectionDefinition = {