+++ src/render/dom/utils/transform.ts
@@ -2,7 +2,7 @@
  * A list of all transformable axes. We'll use this list to generated a version
  * of each axes for each transform.
  */
-export const transformAxes = ["", "X", "Y", "Z"]
+const transformAxes = ["", "X", "Y", "Z"]
 
 /**
  * An ordered array of each transformable value. By default, transform values
@@ -19,9 +19,9 @@ export const boxDistortingKeys: Set<string> = new Set()
  * Generate a list of every possible transform key.
  */
 export const transformProps = ["transformPerspective", "x", "y", "z"]
-order.forEach(operationKey => {
+order.forEach((operationKey) => {
     const isDistorting = new Set(["rotate", "skew"]).has(operationKey)
-    transformAxes.forEach(axesKey => {
+    transformAxes.forEach((axesKey) => {
         const key = operationKey + axesKey
         transformProps.push(key)
         isDistorting && boxDistortingKeys.add(key)